Amend the two numbers, `num1` and `num2`, to be doubles/decimals.

+++ main.c
@@ -3,14 +3,14 @@
 
 int main() {
     /* Declare two numbers */
-    int num1;
-    int num2;
+    double num1;
+    double num2;
     /* Enter numbers */
     printf("Enter first number: ");
-    scanf("%d", &num1);
+    scanf("%lf", &num1);
     printf("Enter second number: ");
-    scanf("%d", &num2);
+    scanf("%lf", &num2);
     /* Print answer (add both numbers together) */
-    printf("Answer: %d", num1 + num2);
+    printf("Answer: %f", num1 + num2);
     return 0;
 }
